% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/argument_convention.R
\name{argument_convention}
\alias{argument_convention}
\title{Default Arguments}
\arguments{
\item{verbose}{(\code{numeric}) \cr verbosity level, incremental;
(0: None, 1: packages that get installed + high-level git operations,
2: includes git checkout infos)}
}
\description{
The documentation to this function lists all the arguments in \code{staged.dependencies}
that are used repeatedly.
}
\keyword{internal}
