% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/install_from_stages.R
\name{check_downstream}
\alias{check_downstream}
\title{Check downstream dependencies}
\usage{
check_downstream(
  target = NULL,
  project = ".",
  downstream_repos = NULL,
  dry_install_and_check = FALSE,
  verbose = 0
)
}
\arguments{
\item{target}{feature branch we want to build, includes fallbacks}

\item{project}{directory of project (for which to restore the dependencies according to target)
must be a git repository; currently checked out branch must be a local branch (not a remote branch)}

\item{dry_install_and_check}{whether to install upstream dependencies and run the checks;
useful to see a dry-run (it however updates the cached repos!)}

\item{verbose}{verbosity level (0: None, 1: more, 2: includes git checkout)}
}
\description{
It installs the downstream dependencies and their upstream dependencies,
and then runs \code{rcmdcheck} (\verb{R CMD check}) on the downstream dependencies.
}
