% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/addin_helpers.R
\name{check_downstream_job}
\alias{check_downstream_job}
\title{Check downstream job}
\usage{
check_downstream_job(project = ".", check_args = Sys.getenv("RCMDCHECK_ARGS"))
}
\arguments{
\item{project}{(\code{character}) directory of project (for which to restore the
dependencies according to feature); must be a git repository.}

\item{check_args}{(\code{list}) arguments passed to \code{rcmdcheck}}
}
\description{
Check downstream job
}
\seealso{
check_downstream
}
