% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dependencies.R
\name{check_yamls_consistent}
\alias{check_yamls_consistent}
\title{Checks that the staged dependency yamls are consistent with
the dependencies listed in the DESCRIPTION files}
\usage{
check_yamls_consistent(dep_structure, skip_if_missing_yaml = TRUE)
}
\arguments{
\item{dep_structure}{\code{dependency_structure} object}

\item{skip_if_missing_yaml}{\code{logical} should checks be skipped on packages
without yaml files. Default \code{TRUE}}
}
\value{
NULL if successful. An error is thrown if inconsistencies found
}
\description{
Checks that the staged dependency yamls are consistent with
the dependencies listed in the DESCRIPTION files
}
\details{
This function explicitly checks that for all packages in the
\code{dependency_structure} object: all upstream and downstream packages specified
in each yaml file are found in the appropriate package DESCRIPTION file
}
\examples{
\dontrun{
x <- dependency_table(project = ".")
check_yamls_consistent(x)
}
}
