% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dependencies.R
\name{dependency_table}
\alias{dependency_table}
\title{TODO}
\usage{
dependency_table(
  project = ".",
  feature = NULL,
  local_repos = get_local_pkgs_from_config(),
  direction = c("upstream", "downstream"),
  verbose = 0
)
}
\arguments{
\item{project}{(`character`) directory of project (for which to restore the
dependencies according to feature); must be a git repository.}

\item{feature}{(`character`) feature we want to build; inferred from the
branch of the project if not provided; warning if not consistent with
current branch of project}
}
\description{
TODO
}
