% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/install_from_stages.R
\name{determine_remote_ref}
\alias{determine_remote_ref}
\title{Implement Staging Rules}
\usage{
determine_remote_ref(
  local_branch,
  remote_branches,
  overall_fallback = "main",
  feature_keyword = "^feature:"
)
}
\arguments{
\item{remote_branches}{names of}

\item{local_brach}{name of branch of package to install}
}
\description{
Implement Staging Rules
}
\examples{

determine_remote_ref("01_abc", c("main", "01_abc"))
determine_remote_ref("01_abc", c("main", "01_abc"), feature_keyword = "") # always match branches

determine_remote_ref("pre-release/01_abc", c("main", "01_abc", "pre-release"))

determine_remote_ref("feature:01_abc", c("main", "feature:01_abc", "devel"))

determine_remote_ref("feature:01_abc", c("main", "feature:01_abc", "devel"))


}
