% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dependencies.R
\name{get_local_pkgs_from_config}
\alias{get_local_pkgs_from_config}
\title{Loads the config file and extracts \code{local_packages}}
\usage{
get_local_pkgs_from_config()
}
\value{
local_packages
}
\description{
Checks that all directories exist and are absolute paths.
}
\examples{
get_local_pkgs_from_config()
}
