% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dependencies_app.R
\name{install_deps_app}
\alias{install_deps_app}
\title{Gadget or Shiny app to select the dependencies to install}
\usage{
install_deps_app(
  default_repo = NULL,
  default_host = "https://github.com",
  default_ref = "main",
  run_gadget = TRUE,
  run_as_job = TRUE,
  verbose = 1,
  install_external_deps = TRUE,
  ...
)
}
\arguments{
\item{default_ref}{(\code{character}) default ref (branch/tag), see also the parameter
\code{ref} of \verb{\link{dependency_table}}. If \code{NULL} this must be entered by user.}

\item{run_gadget}{(\code{logical}) whether to run the app as a gadget}

\item{run_as_job}{(\code{logical}) whether to run the installation as an RStudio job.}

\item{verbose}{verbosity level, incremental; from 0 (none) to 2 (high)}

\item{install_external_deps}{logical to describe whether to install
external dependencies of packages using \code{remotes::install_deps}.}

\item{...}{Additional args passed to \verb{remotes::install_deps. Note }upgrade`
is set to "never" and shouldn't be passed into this function.}

\item{project}{(\code{character}) Should be of the format \code{"<<repo>>@<<host>>"}, for example
\code{"openpharma/stageddeps.water@https://github.com"}, if \code{host} is not included then it
is assumed to be \code{"https://github.com"}. If \code{NULL} this must be entered by user.}
}
\value{
\code{shiny.app} or value returned by app (executed as a gadget)
}
\description{
The dependencies are obtained by traversing the upstream and downstream repositories
in the package's staged dependencies yaml files starting from \code{project}.
}
\examples{
\dontrun{
  install_deps_app("openpharma/stageddeps.food")
}
}
