% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rstudio_jobs.R
\name{install_deps_job}
\alias{install_deps_job}
\title{Install dependencies job}
\usage{
install_deps_job(project = ".", verbose = 1, create_args = list(), ...)
}
\arguments{
\item{project}{(`character`) directory of project (for which to calculate the
dependency structure); must be a git repository.}

\item{verbose}{(`numeric`) verbosity level, incremental;
(0: None, 1: packages that get installed + high-level git operations,
2: includes git checkout infos)}

\item{create_args}{\code{named list} - additional arguments passed to `dependency_table` function}

\item{...}{
  Arguments passed on to \code{\link[=install_deps]{install_deps}}
  \describe{
    \item{\code{dep_structure}}{(\code{dependency_structure}) output of function
\code{dependency_table}}
    \item{\code{install_project}}{(\code{logical}) whether to also install the current
package (\code{project})}
    \item{\code{dry_install}}{(\code{logical}) dry run that lists packages that would be
installed without installing; this still checks out the git repos to
match \code{feature}}
    \item{\code{install_direction}}{"upstream", "downstream" or both; which packages
to install (according to dependency structure)}
    \item{\code{install_external_deps}}{logical to describe whether to install
external dependencies of package using \code{remotes::install_deps}.}
  }}
}
\description{
Install dependencies job
}
\examples{
\dontrun{
install_deps_job()
install_deps_job(create_args = list(feature = "6_makegraph@main"))

# install all dependencies
install_deps_job(create_args = list(direction = c("upstream", "downstream")))
install_deps_job(dry_install = TRUE)
}
}
\seealso{
install_deps
}
