% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/install_from_stages.R
\name{install_repo_add_sha}
\alias{install_repo_add_sha}
\title{Install git repository}
\usage{
install_repo_add_sha(repo_dir)
}
\arguments{
\item{repo_dir}{directory of repo}
}
\description{
It adds the git SHA to the DESCRIPTION file, so that the package
does not need to be installed again when the same commit is already
installed.
}
