% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/install_from_stages.R
\name{install_staged_dependency}
\alias{install_staged_dependency}
\title{Install a package from a git repo with branch corresponding to target}
\usage{
install_staged_dependency(x, target, ...)
}
\arguments{
\item{x}{info about how to fetch the package from git (GitHub or GitLab)}

\item{target}{feature branch we want to build, includes fallbacks}

\item{...}{additional args to pass to installation functions}
}
\description{
Install a package from a git repo with branch corresponding to target
}
\examples{
\dontrun{
install_staged_dependency(list(repo = "openpharma/staged.dependencies", host = "https://api.github.com",
type = "github"), "master", upgrade = FALSE)
install_staged_dependency(list(repo = "NEST/utils.nest", host = "https://github.roche.com/api/v3",
                               type = "github"), "master", upgrade = FALSE)
install_staged_dependency(list(repo = "nest/scda.2021", host = "https://code.roche.com", type = "gitlab"), "master", upgrade = FALSE)

# deps_file <- yaml::read_yaml("staged_dependencies.yaml")
# install_staged_dependency(deps_file[["upstream_repos"]][[3]], "master", upgrade = FALSE)
}

}
