% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/install_from_stages.R
\name{install_upstream_deps}
\alias{install_upstream_deps}
\title{Install upstream dependencies of project corresponding to target}
\usage{
install_upstream_deps(
  target = NULL,
  project = ".",
  install_project = TRUE,
  dry_install = FALSE,
  verbose = FALSE
)
}
\arguments{
\item{target}{feature branch we want to build, includes fallbacks}

\item{project}{directory of project (for which to restore the dependencies according to target)
must be a git repository; currently checked out branch must be a local branch (not a remote branch)}

\item{install_project}{whether to also install the current package (project)}

\item{dry_install}{whether to install or just print (useful for dry-runs, but still
checks out the git repos)}

\item{verbose}{verbosity level (0: None, 1: more, 2: includes git checkout)}
}
\value{
installed packages in installation order
}
\description{
This reads the upstream dependencies for the project and installs the right branches based
on the target.
}
\examples{
\dontrun{
install_upstream_deps()
}

}
