% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/caching.R
\name{rec_checkout_internal_deps}
\alias{rec_checkout_internal_deps}
\title{Recursively check out all repos to match branch determined by ref
starting from repos_to_process.}
\usage{
rec_checkout_internal_deps(
  repos_to_process,
  ref,
  direction = "upstream",
  local_repos = get_local_pkgs_from_config(),
  fallback_branch = "main",
  verbose = 0
)
}
\arguments{
\item{repos_to_process}{\code{list} of \code{list(repo, host)} repos to start from}

\item{ref}{(\code{character}) tag/branch to build}

\item{direction}{(\code{character}) direction in which to discover packages
either or both of "upstream" and "downstream"
to recursively checkout upstream and/or downstream dependencies}

\item{local_repos}{(\code{data.frame}) repositories that should be taken from
local rather than cloned; columns are \verb{repo, host, directory}}

\item{fallback_branch}{(\code{character}) the default branch to try to use if
no other matches found}

\item{verbose}{(\code{numeric}) verbosity level, incremental;
(0: None, 1: packages that get installed + high-level git operations,
2: includes git checkout infos)}
}
\value{
A data frame, one row per checked out repository with columns
repo, host and cache_dir
}
\description{
This uses the \code{staged_dependencies.yaml} to discover the upstream
and downstream packages.
Another function allows to check that only (and all) direct upstream
and downstream packages are listed there.
The packages listed there are internal packages. All other dependencies
listed in the \code{DESCRIPTION} file are external dependencies.
}
