% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/caching.R
\name{rec_checkout_repos}
\alias{rec_checkout_repos}
\title{Recursively check out all repos to match branch determined by feature
starting from repos_to_process.}
\usage{
rec_checkout_repos(
  repos_to_process,
  feature,
  direction = c("upstream"),
  local_repos = get_local_pkgs_from_config(),
  verbose = 0
)
}
\arguments{
\item{repos_to_process}{\code{list} of \code{list(repo, host)}}

\item{feature}{(\code{character}) feature to build}

\item{direction}{(\code{character}) either or both of "upstream" and "downstream"
to recursively checkout upstream and/or downstream dependencies}

\item{local_repos}{(\code{data.frame}) repositories that should be taken from
local rather than cloned; columns are \verb{repo, host, directory}}

\item{verbose}{(\code{numeric}) verbosity level, incremental;
(0: None, 1: packages that get installed + high-level git operations,
2: includes git checkout infos)}
}
\value{
the upstream and/or downstream dependency graphs
}
\description{
Recursively check out all repos to match branch determined by feature
starting from repos_to_process.
}
