% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/install_from_stages.R
\name{restore}
\alias{restore}
\title{Restore Staged Dependencies}
\usage{
restore(
  local_branch = NULL,
  project = ".",
  upgrade = c("default", "ask", "always", "never")
)
}
\description{
Restore Staged Dependencies
}
\examples{
\dontrun{
restore()
}
}
