% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dependencies_helper.R
\name{run_package_actions}
\alias{run_package_actions}
\title{Run an action on the packages}
\usage{
run_package_actions(
  pkg_df,
  action,
  internal_pkg_deps,
  dry = FALSE,
  install_external_deps = TRUE,
  rcmd_args = NULL,
  artifact_dir = NULL,
  verbose = verbose,
  ...
)
}
\arguments{
\item{pkg_df}{table of packages to install, e.g. returned by
`dependency_table(...)$table`}

\item{action}{a subset of "test", "build", "check", "install";
-  if test: run with devtools::test
-  if build: R CMD build
-  if check: `rcmdcheck` with output to command line (if no build action),
   otherwise `R CMD check` on tarball
-  if install: R install based on remotes::install (if no build action),
   otherwise `R CMD INSTALL` on tarball}

\item{internal_pkg_deps}{packages to install (when `install_external_deps = TRUE`)}

\item{dry}{logical, if `FALSE` (default) run the actions, if `TRUE` do not}

\item{install_external_deps}{logical to describe whether to install
external dependencies of package using `remotes::install_deps`.}

\item{rcmd_args}{list with names `build`, `check`,
`install` which are vectors that are passed as separate arguments
to the `R CMD` commands}

\item{artifact_dir}{directory to store log files}

\item{verbose}{verbosity level, incremental - from 0 (none) to 2 (high)}

\item{...}{Additional args passed to `remotes::install_deps. Note `upgrade`
is set to "never" and shouldn't be passed into this function.}
}
\description{
Run an action on the packages
}
