% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dependencies.R
\name{update_stage_dep_yaml_with_direct_deps}
\alias{update_stage_dep_yaml_with_direct_deps}
\title{TODO roxygen}
\usage{
update_stage_dep_yaml_with_direct_deps(
  project = ".",
  feature = NULL,
  local_repos = get_local_pkgs_from_config(),
  verbose = 0
)
}
\description{
TODO roxygen
}
