% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dependencies.R
\name{update_with_direct_deps}
\alias{update_with_direct_deps}
\title{Update existing stage_dependencies yaml file}
\usage{
update_with_direct_deps(dep_structure)
}
\arguments{
\item{dep_structure, }{\code{dep_structure} object, output of \code{dependency_table}
function with \code{project_type = "local"}}
}
\description{
Using the existing stage_dependencies yaml file
'graph' to define internal dependencies, update the
project yaml file to include to include all direct
(i.e. distance 1) upstream and downstream repos
}
