% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{verbose_sd_option}
\alias{verbose_sd_option}
\alias{verbose_sd_set}
\title{Set staged.dependencies verbosity}
\usage{
verbose_sd_set(verbose = 1)
}
\arguments{
\item{verbose}{(\code{numeric}) \cr verbosity level, incremental;
(0: None, 1: packages that get installed + high-level git operations,
2: includes git checkout infos)}
}
\description{
Functions to set and remove the option parameter \code{verbose_level_staged.deps}.
It can assume integer values between \code{c(0, 1, 2)}. This will set this variable
as an option with \code{\link[=options]{options()}} and \code{\link[=getOption]{getOption()}}.
}
\examples{
verbose_sd_set(2)
verbose_sd_get() # 2, the inserted value
verbose_sd_rm()
verbose_sd_get() # 1, the default

}
