% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dependencies.R
\name{build_check_install}
\alias{build_check_install}
\title{Build, check and install internal dependencies}
\usage{
build_check_install(
  dep_structure,
  install_direction = c("upstream", "downstream"),
  packages_to_process = NULL,
  verbose = 0,
  steps = c("build", "check", "install"),
  rcmd_args = list(check = c("--no-manual")),
  artifact_dir = tempfile(),
  install_external_deps = TRUE,
  ...
)
}
\arguments{
\item{dep_structure}{(\code{dependency_structure}) output of function
\code{dependency_table}}

\item{install_direction}{"upstream", "downstream" or both; which packages
to install (according to dependency structure). By default this is only "upstream"}

\item{packages_to_process}{(\code{character}) An additional filter, only packages on
this list will be installed (advanced usage only)}

\item{verbose}{verbosity level, incremental; from 0 (none) to 2 (high)}

\item{steps}{(\code{character} vector) subset of "build", "check", "install";
useful to skip checking for example}

\item{rcmd_args}{(\code{list}) with names \code{build}, \code{check},
\code{install} which are vectors that are passed as separate arguments
to the \verb{R CMD} commands}

\item{artifact_dir}{(\code{character}) directory to place built R packages
and logs}

\item{install_external_deps}{logical to describe whether to install
external dependencies of packages using \code{remotes::install_deps}.}

\item{...}{
  Arguments passed on to \code{\link[=install_deps]{install_deps}}
  \describe{
    \item{\code{install_project}}{(\code{logical}) whether to also install the current
package (i.e. the package named in \code{dependency_structure$current_pkg})}
    \item{\code{dry_install}}{(\code{logical}) dry run that lists packages that would be
installed}
    \item{\code{dependency_packages}}{(\code{character}) An additional filter, only packages on this
list will be installed (advanced usage only)}
  }}
}
\value{
\code{artifact_dir} directory with log files
}
\description{
Build, check and install internal dependencies
}
\examples{
\dontrun{
  x <- dependency_table(project = ".", verbose = 1)
  build_check_install(x, steps = c("build", "check"), verbose = 1)
  build_check_install(x, artifact_dir = "../output")

}
}
