% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/caching.R
\name{clear_cache}
\alias{clear_cache}
\title{Clear the repository cache}
\usage{
clear_cache(pattern = "*")
}
\arguments{
\item{pattern}{files to remove, see \code{unlink} (wildcards \code{*} and \verb{?} allowed)}
}
\description{
Use this function to cleanly clone some repositories if the
\code{git} operations fail.
}
