% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dependencies.R
\name{dependency_table}
\alias{dependency_table}
\title{Get the dependency structure as a table}
\usage{
dependency_table(
  project = ".",
  feature = NULL,
  local_repos = get_local_pkgs_from_config(),
  direction = c("upstream", "downstream"),
  verbose = 0
)
}
\arguments{
\item{project}{(\code{character}) directory of project (for which to restore the
dependencies according to feature); must be a git repository.}

\item{feature}{(\code{character}) feature we want to build; inferred from the
branch of the project if not provided; warning if not consistent with
current branch of project}

\item{local_repos}{(\code{data.frame}) repositories that should be taken from
local rather than cloned; columns are \verb{repo, host, directory}}

\item{direction}{(\code{character}) either or both of "upstream" and "downstream"
to recursively checkout upstream and/or downstream dependencies}

\item{verbose}{(\code{numeric}) verbosity level, incremental;
(0: None, 1: packages that get installed + high-level git operations,
2: includes git checkout infos)}
}
\value{
depending on \code{return_table_only} (see above); either \code{data.frame} or a list with
the following elements:

\code{df}:
\code{data.frame} with columns \verb{repo, host, type, branch}, where \code{type} is:
\itemize{
\item \code{current} (project),
\item \code{upstream} (of project),
\item \code{downstream} (of project),
\item \code{other} (the remaining, e.g. downstream dependencies of upstream dependencies)
}

and \code{branch} is the branch corresponding to \code{feature} and becomes \code{local (branch)}
if package is in \code{local_repos}.

\code{graph}:
visNetwork graph with arrows going in the direction from downstream to upstream.
Arrows are solid if both upstream and downstream list each other; they
are dashed if either does not list the other:
\itemize{
\item if upstream does not list downstream, blue dashed
\item if downstream does not list upstream, red dashed (this means that
the downstream repo's yaml needs to be fixed)
Tooltips give more information about nodes and edges.
}

\code{deps}:
deps: upstream and downstream dependencies of each node
}
\description{
The dependency structure depends on \code{feature} which determines which
repositories are checked out.
It starts from the project and discovers all reachable upstream and
downstream dependencies.
}
\examples{
\dontrun{
dependency_table()

dependency_table(direction = "upstream")
}
}
