% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dependencies.R
\name{get_all_external_dependencies}
\alias{get_all_external_dependencies}
\title{List the external R packages required to be installed}
\usage{
get_all_external_dependencies(
  dep_structure,
  available_packages = as.data.frame(utils::available.packages()),
  install_direction = "upstream",
  package_list = NULL,
  from_internal_dependencies = c("Depends", "Imports", "LinkingTo", "Suggests"),
  from_external_dependencies = c("Depends", "Imports", "LinkingTo")
)
}
\arguments{
\item{dep_structure}{(\code{dependency_structure}) output of function
\code{dependency_table}; uses \code{dep_structure$table} to infer the packages
to apply action to and infer installation order;
uses \code{dep_structure$deps} to infer upstream dependencies}

\item{available_packages}{(\code{data.frame}) A dataframe of the format given by
\code{as.data.frame(utils::available.packages)}. It is unlikely this default needs to be changed;
however you need to ensure the \code{options("repos")} contains the urls of all expected repos
(e.g. Bioconductor).}

\item{install_direction}{"upstream", "downstream" or "all"; which packages
to install (according to dependency structure). By default this is only "upstream"}

\item{package_list}{(\code{character}) If not NULL, an additional filter, only packages on this
list will be considered and their dependencies installed if needed (advanced usage only).}

\item{from_internal_dependencies}{Vector chosen from \code{c("Depends", "Imports", "LinkingTo", "Suggests", "Enhances")}
which fields of the DESCRIPTION file of the internal packages should be included. Default:
\code{c("Depends", "Imports", "LinkingTo", "Suggests")}}

\item{from_external_dependencies}{Vector chosen from \code{c("Depends", "Imports", "LinkingTo", "Suggests", "Enhances")}
which fields of the DESCRIPTION file of the internal packages should be included. Default:
\code{c("Depends", "Imports", "LinkingTo")}}
}
\value{
A vector of 'external' R packages required to install
the selected 'internal' packages, ordered by install order (unless \code{from_external_dependencies}
does not include \code{"Depends"}, \code{"Imports"} and \code{"LinkingTo"}). The core R packages
(e.g. \code{methods}, \code{utils}) are not included. The output can be used with \code{remotes::system_requirements}
to extract the system requirements needed for your packages, see example below.
}
\description{
List the external R packages required to be installed
}
\examples{
\dontrun{
  x <- dependency_table("openpharma/stageddeps.electricity",
    project_type = "repo@host", feature = "main")

  # get external package dependencies
  ex_deps <- get_all_external_dependencies(x)
  print(ex_deps)

  # get system dependencies (in this case there are none)
  unique(unlist(lapply(ex_deps,
    function(pkg, ...)
      remotes::system_requirements(package = pkg, ...),
    os = "ubuntu",
    os_release = "20.04")
 ))
}
}
