% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dependencies_app.R
\name{install_deps_app}
\alias{install_deps_app}
\title{Gadget or Shiny app to select the dependencies to install}
\usage{
install_deps_app(
  default_repo = NULL,
  default_host = "https://github.com",
  default_ref = "main",
  fallback_branch = "main",
  run_gadget = TRUE,
  run_as_job = TRUE,
  verbose = 1,
  install_external_deps = TRUE,
  renv_profile = NULL,
  upgrade = "never",
  ...
)
}
\arguments{
\item{default_repo}{(\code{character}) the repository name for the dependency graph to be
created for, for example, \code{"openpharma/stageddeps.water"}.
If \code{NULL} this must be entered by app user and can always be changed by the user.}

\item{default_host}{(\code{character}) the host for the repository for the dependency graph to
be created for by default \code{"https://github.com"}. If \code{NULL} this must be entered by app
user and can always be changed by the user.}

\item{default_ref}{(\code{character}) default ref (branch/tag), see also the parameter
\code{ref} of \verb{\link{dependency_table}}. If \code{NULL} this must be entered by app user
and can always be changed by the user.}

\item{fallback_branch}{(\code{character}) the default branch to try to use if
no other matches found}

\item{run_gadget}{(\code{logical}) whether to run the app as a gadget}

\item{run_as_job}{(\code{logical}) whether to run the installation as an RStudio job.}

\item{verbose}{verbosity level, incremental; from 0 (none) to 2 (high)}

\item{install_external_deps}{logical to describe whether to install
external dependencies of package using \code{\link[remotes:install_deps]{remotes::install_deps()}} (or \code{\link[renv:install]{renv::install()}} if
inside an renv environment) .}

\item{renv_profile}{(\code{character}) the name of the \code{renv} profile of the \code{renv.lock} files
to be included from the repos. The standard \code{renv.lock} file uses the default \code{NULL} argument here.}

\item{upgrade}{argument passed to \code{\link[remotes:install_deps]{remotes::install_deps()}}, defaults to \code{'never'}. Ignored
if inside an \code{renv} environment.}

\item{...}{Additional args passed to \code{\link[remotes:install_deps]{remotes::install_deps()}}. Ignored
if inside an \code{renv} environment.}
}
\value{
\code{shiny.app} or value returned by app (executed as a gadget)
}
\description{
The dependencies are obtained by traversing the upstream and downstream repositories
in the package's staged dependencies yaml files starting from \code{project}.
}
\examples{
\dontrun{
  install_deps_app("openpharma/stageddeps.food")
}
}
