% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dependencies.R
\name{install_deps_app}
\alias{install_deps_app}
\title{Gadget or Shiny app to select the dependencies to install}
\usage{
install_deps_app(
  project = ".",
  feature = NULL,
  local_repos = get_local_pkgs_from_config(),
  run_gadget = TRUE,
  verbose = 1
)
}
\arguments{
\item{project}{(\code{character}) directory of project (for which to restore the
dependencies according to feature); must be a git repository.}

\item{feature}{(\code{character}) feature we want to build; inferred from the
branch of the project if not provided; warning if not consistent with
current branch of project}

\item{local_repos}{(\code{data.frame}) repositories that should be taken from
local rather than cloned; columns are \verb{repo, host, directory}}

\item{run_gadget}{whether to run the app as a gadget}

\item{verbose}{(\code{numeric}) verbosity level, incremental;
(0: None, 1: packages that get installed + high-level git operations,
2: includes git checkout infos)}
}
\value{
\code{shiny.app} or value returned by app (executed as a gadget)
}
\description{
The dependencies are obtained by traversing the upstream and downstream
dependencies starting from \code{project}.
}
