% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dependencies_app.R
\name{install_deps_app}
\alias{install_deps_app}
\title{Gadget or Shiny app to select the dependencies to install}
\usage{
install_deps_app(
  project = ".",
  default_ref = NULL,
  local_repos = get_local_pkgs_from_config(),
  run_gadget = TRUE,
  run_as_job = TRUE,
  verbose = 1,
  install_external_deps = TRUE,
  upgrade = "never",
  ...
)
}
\arguments{
\item{project}{(\code{character}) directory of project (for which to calculate the
dependency structure); must be a git repository.}

\item{default_ref}{(\code{character}) default ref (branch/tag), see also the parameter
\code{ref} of \verb{\link{dependency_table}}}

\item{local_repos}{(\code{data.frame}) repositories that should be taken from
local rather than cloned; columns are \verb{repo, host, directory}}

\item{run_gadget}{(\code{logical}) whether to run the app as a gadget}

\item{run_as_job}{(\code{logical}) whether to run the installation as an RStudio job.}

\item{verbose}{verbosity level, incremental; from 0 (none) to 2 (high)}

\item{install_external_deps}{logical to describe whether to install
external dependencies of packages using \code{remotes::install_deps}.}

\item{upgrade}{argument passed to \code{remotes::install_deps}, defaults to 'never'.}

\item{...}{Additional args passed to `remotes::install_deps.}
}
\value{
\code{shiny.app} or value returned by app (executed as a gadget)
}
\description{
The dependencies are obtained by traversing the upstream and downstream repositories
in the package's staged dependencies yaml files starting from \code{project}.
}
