% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dependencies.R
\name{install_deps_app}
\alias{install_deps_app}
\title{Gadget or Shiny app to select the dependencies to install}
\usage{
install_deps_app(
  project = ".",
  default_feature = NULL,
  local_repos = get_local_pkgs_from_config(),
  run_gadget = TRUE,
  run_as_job = TRUE,
  verbose = 1,
  install_external_deps = TRUE,
  ...
)
}
\arguments{
\item{project}{(\code{character}) directory of project (for which to restore the
dependencies according to feature); must be a git repository.}

\item{default_feature}{(\code{character}) default feature, see also the parameter
\code{feature} of \verb{\link{install_deps_project}}}

\item{local_repos}{(\code{data.frame}) repositories that should be taken from
local rather than cloned; columns are \verb{repo, host, directory}}

\item{run_gadget}{(\code{logical}) whether to run the app as a gadget}

\item{run_as_job}{(\code{logical}) whether to run the installation as an RStudio job.}

\item{verbose}{(\code{numeric}) verbosity level, incremental;
(0: None, 1: packages that get installed + high-level git operations,
2: includes git checkout infos)}

\item{install_external_deps}{logical to describe whether to install
external dependencies of package using \code{remotes::install_deps}.}

\item{...}{Additional args passed to \verb{remotes::install_deps. Note }upgrade`
is set to "never" and shouldn't be passed into this function.}
}
\value{
\code{shiny.app} or value returned by app (executed as a gadget)
}
\description{
The dependencies are obtained by traversing the upstream and downstream
dependencies starting from \code{project}.
}
