% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rstudio_jobs.R
\name{install_deps_job}
\alias{install_deps_job}
\title{Install dependencies job}
\usage{
install_deps_job(
  project = ".",
  project_type = "local",
  verbose = 1,
  create_args = list(renv_profile = Sys.getenv("RENV_PROFILE")),
  ...
)
}
\arguments{
\item{project}{(\code{character}) If \code{project_type} is \code{local} then
directory of project (for which to calculate the dependency structure);
must be a git repository. If \code{project_type} is \code{repo@host} then should
be character of the form \verb{openpharma/stageddeps.food@https://github.com}
If host is not included in the string then the default \verb{https://github.com}
is assumed.}

\item{project_type}{(\code{character}) See \code{project} argument.}

\item{verbose}{(\code{numeric}) \cr verbosity level, incremental;
(0: None, 1: packages that get installed + high-level git operations,
2: includes git checkout infos)}

\item{create_args}{\code{named list} - additional arguments passed to \code{dependency_table} function}

\item{...}{
  Arguments passed on to \code{\link[=install_deps]{install_deps}}
  \describe{
    \item{\code{dep_structure}}{(\code{dependency_structure}) output of function
\code{dependency_table}; uses \code{dep_structure$table} to infer the packages
to apply action to and infer installation order;
uses \code{dep_structure$deps} to infer upstream dependencies}
    \item{\code{install_project}}{(\code{logical}) whether to also install the current
package (i.e. the package named in \code{dependency_structure$current_pkg}),
ignored unless \code{install_direction = "upstream"} (because downstream
deps automatically install all their upstream deps)}
    \item{\code{install_direction}}{"upstream", "downstream" or "all"; which packages
to install (according to dependency structure). By default this is only "upstream"}
    \item{\code{install_external_deps}}{logical to describe whether to install
external dependencies of package using \code{\link[remotes:install_deps]{remotes::install_deps()}} (or \code{\link[renv:install]{renv::install()}} if
inside an renv environment) .}
    \item{\code{upgrade}}{argument passed to \code{\link[remotes:install_deps]{remotes::install_deps()}}, defaults to \code{'never'}. Ignored
if inside an \code{renv} environment.}
    \item{\code{package_list}}{(\code{character}) If not NULL, an additional filter, only packages on this
list will be considered and their dependencies installed if needed (advanced usage only).}
    \item{\code{dry}}{(\code{logical}) dry run that outputs what would happen without actually
doing it.}
  }}
}
\description{
Install dependencies job
}
\examples{
\dontrun{
install_deps_job()
install_deps_job(create_args = list(ref = "6_makegraph@main"))

# install all dependencies
install_deps_job(create_args = list(direction = "all"))
install_deps_job(dry_install = TRUE)
}
}
\seealso{
install_deps
}
