% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/git_tools.R
\name{install_repo_add_sha}
\alias{install_repo_add_sha}
\title{Install a git repository}
\usage{
install_repo_add_sha(repo_dir, ...)
}
\arguments{
\item{repo_dir}{directory of repo}

\item{...}{Additional args passed to \code{remotes::install_deps}. Note \code{upgrade}
is set to "never" and shouldn't be passed into this function.}
}
\description{
It adds the git SHA to the DESCRIPTION file, so that the package
does not need to be installed again when the same commit is already
installed.
}
