% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/caching.R
\name{rec_checkout_internal_deps}
\alias{rec_checkout_internal_deps}
\title{Recursively check out all repos to match branch determined by feature
starting from repos_to_process.}
\usage{
rec_checkout_internal_deps(
  repos_to_process,
  feature,
  direction = c("upstream"),
  local_repos = get_local_pkgs_from_config(),
  verbose = 0
)
}
\arguments{
\item{repos_to_process}{\code{list} of \code{list(repo, host)}}

\item{feature}{(\code{character}) feature to build}

\item{direction}{(\code{character}) either or both of "upstream" and "downstream"
to recursively checkout upstream and/or downstream dependencies}

\item{local_repos}{(\code{data.frame}) repositories that should be taken from
local rather than cloned; columns are \verb{repo, host, directory}}

\item{verbose}{(\code{numeric}) verbosity level, incremental;
(0: None, 1: packages that get installed + high-level git operations,
2: includes git checkout infos)}
}
\value{
A list, on entry per checkout out repository whose name is
the name of the repo in the form \code{repo @ host} (\code{hash(repo, host)})
and whose value is the directory to which it has been checked out into
}
\description{
This uses the \code{staged_dependencies.yaml} to discover the upstream
and downstream packages.
Another function allows to check that only (and all) direct upstream
and downstream packages are listed there.
The packages listed there are internal packages. All other dependencies
listed in the \code{DESCRIPTION} file are external dependencies.
}
