% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/install_from_stages.R
\name{restore}
\alias{restore}
\title{Restore Staged Dependencies}
\usage{
restore(target = NULL, project = ".", ...)
}
\arguments{
\item{target}{feature branch we want to build, includes fallbacks}

\item{...}{additional args to pass to installation functions}
}
\description{
Restore Staged Dependencies
}
\examples{
\dontrun{
restore()
}
}
