% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dependencies_helper.R
\name{run_package_actions}
\alias{run_package_actions}
\title{Run an action on the packages}
\usage{
run_package_actions(
  pkg_df,
  action,
  internal_pkg_deps,
  dry = FALSE,
  install_external_deps = TRUE,
  rcmd_args = NULL,
  artifact_dir = NULL,
  verbose = verbose,
  ...
)
}
\arguments{
\item{pkg_df}{table of packages to install, e.g. returned by
`dependency_table(...)$table`}

\item{action}{a subset of "test", "build", "check", "install";
-  if test: run with devtools::test
-  if build: R CMD build
-  if check: `rcmdcheck` with output to command line (if no build action),
   otherwise `R CMD check` on tarball
-  if install: R install based on remotes::install (if no build action),
   otherwise `R CMD INSTALL` on tarball}

\item{internal_pkg_deps}{packages to install (when `install_external_deps = TRUE`)}

\item{install_external_deps}{logical to describe whether to install
external dependencies of package using `remotes::install_deps`.}
}
\description{
Run an action on the packages
}
