% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dependencies_helper.R
\name{run_package_actions}
\alias{run_package_actions}
\title{Run an action on the packages}
\usage{
run_package_actions(
  pkg_actions,
  internal_pkg_deps,
  dry = FALSE,
  install_external_deps = TRUE,
  upgrade = "never",
  rcmd_args = NULL,
  artifact_dir = NULL,
  verbose = 0,
  ...
)
}
\arguments{
\item{pkg_actions}{sorted \code{data.frame} with columns \code{cache_dir} pointing
to package source, \code{sha}, \code{installable} and \code{actions}, processed from top to bottom in order
Each \code{actions} is a subset of "test", "build", "check", "install";
\itemize{
\item if test: run with \code{devtools::test}
\item if build: \verb{R CMD build}
\item if check: \code{rcmdcheck} with output to command line (if no build action),
otherwise \verb{R CMD check} on tarball
\item if install: R install based on \code{remotes::install} (if no build action),
otherwise \verb{R CMD INSTALL} on tarball
}}

\item{internal_pkg_deps}{packages to not install (when \code{install_external_deps = TRUE})}

\item{dry}{logical, if \code{FALSE} (default) run the actions, if \code{TRUE} do not}

\item{install_external_deps}{logical to describe whether to install
external dependencies of package using \code{remotes::install_deps} (or \code{renv::install} if
inside an renv environment) .}

\item{upgrade}{argument passed to \code{remotes::install_deps}, defaults to 'never'. Ignored
if inside an \code{renv} environment.}

\item{rcmd_args}{list with names \code{build}, \code{check},
\code{install} which are vectors that are passed as separate arguments
to the \verb{R CMD} commands}

\item{artifact_dir}{directory to store log files, only when actions include
\code{build}; action \code{test} only outputs to the console}

\item{verbose}{verbosity level, incremental - from 0 (none) to 2 (high)}

\item{...}{Additional args passed to \verb{remotes::install_deps Ignored if inside an }renv` environment.}
}
\description{
Run an action on the packages
}
