% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/install_from_stages.R
\name{topological_sort}
\alias{topological_sort}
\title{Topologically sorts nodes so that parents are listed before all their children}
\usage{
topological_sort(child_to_parents)
}
\arguments{
\item{child_to_parents}{mapping from child to its parents (upstream dependencies)}
}
\value{
vector listing parents before children
}
\description{
Topologically sorts nodes so that parents are listed before all their children
}
\examples{
topological_sort <- staged.dependencies:::topological_sort

all(topological_sort(list(
n1 = c(), n2 = c("n1"), n3 = c("n2"),
n4 = c("n2", "n3"))) == c("n1", "n2", "n3", "n4")
)
is.null(topological_sort(list()))

}
