% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dependencies.R
\name{update_stageddeps_yaml_with_direct_deps}
\alias{update_stageddeps_yaml_with_direct_deps}
\title{Update existing stage_dependencies yaml file}
\usage{
update_stageddeps_yaml_with_direct_deps(
  project = ".",
  feature = NULL,
  local_repos = get_local_pkgs_from_config(),
  verbose = 0
)
}
\arguments{
\item{project}{(\code{character}) directory of project (for which to restore the
dependencies according to feature); must be a git repository.}

\item{feature}{(\code{character}) feature we want to build; inferred from the
branch of the project if not provided; warning if not consistent with
current branch of project}

\item{local_repos}{(\code{data.frame}) repositories that should be taken from
local rather than cloned; columns are \verb{repo, host, directory}}

\item{verbose}{(\code{numeric}) verbosity level, incremental;
(0: None, 1: packages that get installed + high-level git operations,
2: includes git checkout infos)}
}
\description{
Using the existing stage_dependencies yaml file
'graph' to define internal dependencies, update the
project yaml file to include to include all direct
(i.e. distance 1) upstream and downstream repos
}
